requirements.txt
texttoowo
discord.py
asyncio
async-timeout